# Make simulate iterate by card instead of by day.