--- a/src/optimal_retention.rs
+++ b/src/optimal_retention.rs
@@ -232,7 +232,7 @@ pub fn simulate(
             {
                 today += 1;
             }
-            if today + 1 >= learn_span {
+            if today >= learn_span {
                 break;
             }