Simulating the reviews iterating card by card instead of day by day speeds up the calculation (especially with a long learn_span) and simplifies the code.
Also fixes the fact that there's a drop-off near the start.

I hope that this functions in a way pretty much exactly like the day by day version, I'll leave the tests failing so you can see exactly how the results changed.

```js
thread 'optimal_retention::tests::simulate_with_learn_review_limit' panicked at src/optimal_retention.rs:909:9:
assertion `left == right` failed
  left: [28, 21, 40, 80, 72, 81, 86, 102, 72, 93, 105, 115, 120, 118, 124, 126, 137, 136, 158, 150, 159, 171, 156, 170, 183, 158, 195, 173, 194, 178]
 right: [0, 15, 16, 27, 69, 66, 81, 83, 75, 85, 90, 102, 107, 105, 122, 120, 141, 117, 132, 153, 153, 154, 152, 152, 149, 157, 166, 165, 184, 174]
```
```js
thread 'optimal_retention::tests::simulator' panicked at src/optimal_retention.rs:864:9:
assertion `left == right` failed
  left: 7778.458
 right: 6521.068
```
```js
thread 'optimal_retention::tests::optimal_retention' panicked at src/optimal_retention.rs:976:9:
assertion `left == right` failed
  left: 0.86305356
 right: 0.8277919
``` 
(/\ I've probably screwed up somewhere)
```js
thread 'optimal_retention::tests::optimal_retention_with_old_parameters' panicked at src/optimal_retention.rs:996:9:
assertion `left == right` failed
  left: 0.8716841
 right: 0.85450846
```
Please note that my results are left shifted one value compared to before because "today" will have a very small cost if the user has already done their reviews. (I think).

And here's how the graphs look with my decks:

## Large deck (Done reviews)
### This pr
![new](https://github.com/user-attachments/assets/36627690-9398-440e-8a8e-d9e74ede7f07)
### Before
(Starting drop-off lasts more than one day so not just the fact reviews are done, if I had to guess the first review of a card isn't registered for some reason)
![old](https://github.com/user-attachments/assets/4d1b4fe3-8079-4191-ac31-702f66eaa369)
## Smaller deck (Undone reviews)
### This pr
![image](https://github.com/user-attachments/assets/2bfbd492-d0f0-4a41-b2d7-55bbab7b4a1c)
### Before
![image](https://github.com/user-attachments/assets/da07d819-f761-4f28-96f6-49bc2b587965)
## Smaller deck (Done reviews)
### This pr
![image](https://github.com/user-attachments/assets/cedbd3b9-c6d8-40f9-bc66-b14a2365058a)
### Before
![image](https://github.com/user-attachments/assets/b40a36ac-fe7d-4616-8813-217195d82620)