@L-M-Sherlock 2024-10-11
It's a good idea to make simulate iterate by card. However, it also makes it impossible to introduce any review sort order into simulation, doesn't it?