@Luc-Mcgrady 2024-10-11
Good point. Hopefully by using a priority queue and iterating through the cards in the order that they are reviewed, changing the review priorities will be possible.