@L-M-Sherlock 2024-10-12
It's impossible that the number of reviews in the first day is larger than zero without any existing cards.

```
---- optimal_retention::tests::simulate_with_learn_review_limit stdout ----
thread 'optimal_retention::tests::simulate_with_learn_review_limit' panicked at src/optimal_retention.rs:931:9:
assertion `left == right` failed
  left: [15, 28, 28, 45, 91, 86, 109, 99, 105, 101, 112, 121, 126, 116, 129, 144, 162, 145, 140, 168, 158, 174, 166, 167, 172, 174, 172, 197, 195, 194]
 right: [0, 15, 16, 27, 69, 66, 81, 83, 75, 85, 90, 102, 107, 105, 122, 120, 141, 117, 132, 153, 153, 154, 152, 152, 149, 157, 166, 165, 184, 174]
```