@Luc-Mcgrady 2024-10-14
I think that's better? I think it was treating new cards as review cards at certain points.
```
 left: [0, 13, 23, 30, 58, 81, 83, 81, 86, 88, 88, 106, 120, 110, 122, 133, 132, 121, 131, 143, 161, 188, 142, 179, 145, 156, 172, 191, 174, 165]
 right: [0, 15, 16, 27, 69, 66, 81, 83, 75, 85, 90, 102, 107, 105, 122, 120, 141, 117, 132, 153, 153, 154, 152, 152, 149, 157, 166, 165, 184, 174]
```