@L-M-Sherlock 2024-10-15
> Simulating the reviews iterating card by card instead of day by day speeds up the calculation (especially with a long learn_span) and simplifies the code.

Could you benchmark it?