@Luc-Mcgrady 2024-10-15
Well on my computer at least:

Current:
```
optimal_retention       time:   [274.85 ms 278.54 ms 282.14 ms]
```
This pr:
```
optimal_retention       time:   [35.805 ms 36.197 ms 36.589 ms]
                        change: [-87.228% -87.005% -86.782%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
```