@L-M-Sherlock 2024-10-15
```
[src\optimal_retention.rs:505:9] iter = 5
thread 'optimal_retention::tests::optimal_retention_with_old_parameters' panicked at src\optimal_retention.rs:1047:9:
assertion `left == right` failed
  left: 0.82624805
 right: 0.85450846
```

This difference is still larger than my expectation. I plan to do a T-test tomorrow to check whether the difference is statistically significant. If not, I will merge it.