@L-M-Sherlock 2024-10-16
After increasing the `sample_size`, the results are close enough. I will test it in Anki for final resolution.