@asukaminato0721 2024-10-16
~~just want to say that rust has built-in priority queue https://doc.rust-lang.org/std/collections/binary_heap/index.html~~ , but currently this works well, so let it be.